+++ BankATMApp/RegularAccount.cs
@@ -11,12 +11,12 @@ namespace BankATMApp
         {
         }
 
-        public override void Credit(int paramAmount)
+        public override void Credit(decimal paramAmount)
         {
             base.Funds += paramAmount;
         }
 
-        public override void Debit(int paramAmount)
+        public override void Debit(decimal paramAmount)
         {
             if (disallowedNegative && base.Funds <= 0 && paramAmount > base.Funds)
             {
